Add files via upload
I am passionate about interior design and furnishing spaces. I believe that furniture brings changes to a space and defines its functionality. The variety of design, range, and type of furniture can transform a dull area into a vibrant experience.

Furniture in our homes holds significant importance for several reasons:

It adds beauty to the house - The wide variety of furniture, from chairs, tables, beds, and sofas, can change the overall look of your home. The grace and beauty of a house are complete with the placement of furniture as part of the decor or utility.

It gives meaning to decor - The presence of furniture makes our home a functional unit. The generous range and additional variety of each category of furniture provide a purpose for the decor (relaxation, cooking, family activities) and make the space well-equipped.

It provides balance - Furniture gives us a balanced view of the entire space in our home. The volumes, shapes, and heights improve our vision of the house, as well as the experiences within it.

It defines the flow of space - Through the placement of furniture, the house is divided into categories of spaces, thus defining zones. Furniture occupies most of the space in the house and helps us constantly bring life back to our home.

For these reasons, I have chosen to develop this concept of a bunk bed that can have a unique utility.